# perf: sonic boom