--- a/benchmarks/logger/src/ogma-json.logger.ts
+++ b/benchmarks/logger/src/ogma-json.logger.ts
@@ -1,0 +1,11 @@
+import { OgmaStream } from '@ogma/common';
+import { Ogma } from '@ogma/logger';
+
+export function createOgmaJsonLogger(stream: OgmaStream) {
+  return new Ogma({
+    application: 'Ogma Bench',
+    stream,
+    logLevel: 'INFO',
+    json: true,
+  });
+}