--- a/packages/common/src/log.interface.ts
+++ b/packages/common/src/log.interface.ts
@@ -1,7 +1,7 @@
 import { OgmaWritableLevel } from './level.enum';
 
 export interface OgmaLog {
-  time: string;
+  time: string | number;
   hostname?: string;
   application?: string;
   context?: string;