Hello again!

This time I bring more performance improvements based on my study of PinoJS, we went from 102k in this https://github.com/jmcdo29/ogma/pull/1491 to:

```
'NestJS Logger x 6,598 ops/sec ±5.64% (80 runs sampled)',
'Ogma x 419,803 ops/sec ±5.68% (76 runs sampled)',
'Pino x 743,735 ops/sec ±2.15% (82 runs sampled)',
'console.log x 140,086 ops/sec ±1.25% (89 runs sampled)',
'process.stdout x 288,950 ops/sec ±2.77% (75 runs sampled)'
```

Now, the profiler information is almost empty hahaha:

```
Statistical profiling result from perf-ogma-stress-v4.log, (4072 ticks, 0 unaccounted, 0 excluded).

 [Shared libraries]:
   ticks  total  nonlib   name
   1659   40.7%          /home/h4ad/.asdf/installs/nodejs/16.16.0/bin/node
    408   10.0%          /usr/lib/x86_64-linux-gnu/libc-2.31.so
     30    0.7%          [vdso]

 [JavaScript]:
   ticks  total  nonlib   name
     44    1.1%    2.2%  LazyCompile: *formatStream /home/h4ad/Projects/opensource/performance-analysis/ogma-performance-analysis/node_modules/@ogma/logger/src/logger/ogma.js:168:17
     26    0.6%    1.3%  LazyCompile: *<anonymous> /home/h4ad/Projects/opensource/performance-analysis/ogma-performance-analysis/dist/ogma-log.stress.js:1:1
      3    0.1%    0.2%  Function: ^SonicBoom.release /home/h4ad/Projects/opensource/performance-analysis/ogma-performance-analysis/node_modules/sonic-boom/index.js:127:18
      1    0.0%    0.1%  LazyCompile: *write node:fs:802:15
      1    0.0%    0.1%  LazyCompile: *normalizeString node:path:66:25
      1    0.0%    0.1%  LazyCompile: *info /home/h4ad/Projects/opensource/performance-analysis/ogma-performance-analysis/node_modules/@ogma/logger/src/logger/ogma.js:238:9
      1    0.0%    0.1%  LazyCompile: *SonicBoom.release /home/h4ad/Projects/opensource/performance-analysis/ogma-performance-analysis/node_modules/sonic-boom/index.js:127:18
      1    0.0%    0.1%  Function: ^wrapper node:fs:803:19
      1    0.0%    0.1%  Function: ^normalizeString node:path:66:25
```

> The number of ticks is bigger, but the time to perform the logs is now much smaller than in the previous version.

> [Source](https://github.com/jmcdo29/ogma/files/10388229/perf-ogma-stress-v4.txt)

The first optimization was the usage of [sonic-boom](https://www.npmjs.com/package/sonic-boom), this library was responsible for more of the performance improvements.

Then, I did a breaking change in `styler` by moving from `getters` to use methods because is way faster.

Finally, the last optimization was inside `ogma.ts`, I almost cache all calls for `colorize`, we had a bunch of places where the value didn't change so we don't need to calculate every time we log something and I also did a happy path in the code to single messages to avoid perform `regex.test`.