@H4ad 2023-01-12
With the latest changes, the benchmark looks like:

```
  'NestJS Logger x 6,797 ops/sec ±5.41% (80 runs sampled)',
  'Ogma x 497,586 ops/sec ±4.41% (78 runs sampled)',
  'Pino x 741,796 ops/sec ±1.29% (88 runs sampled)',
  'console.log x 136,191 ops/sec ±1.83% (91 runs sampled)',
  'process.stdout x 294,599 ops/sec ±1.52% (67 runs sampled)'

  'NestJS Logger x 6,816 ops/sec ±5.62% (83 runs sampled)',
  'Ogma x 500,222 ops/sec ±6.06% (66 runs sampled)',
  'Pino x 719,338 ops/sec ±2.24% (82 runs sampled)',
  'console.log x 134,167 ops/sec ±2.01% (93 runs sampled)',
  'process.stdout x 304,896 ops/sec ±2.18% (71 runs sampled)'
```