@jmcdo29 2023-01-13
I'm gonna play around with the benchmarks and improvements just a little more before merging this in. Trying to get the JSON logger up to the same kind of speed. Thanks for getting this started though! Definitely been a fun learning process