@jmcdo29 2023-01-19
I'm gonna make a couple of changesets for this to document all of the changes in an easy to digest manner. Once a PR for the nest core repo gets merged in with some extra websocket support I can release the next version of `@ogma/nestjs-module` and then I can make all of the breaking changes I'm going to at once. There's a lot of them in the pipeline now and I'd rather just do it all in one fell swoop, which is why I'm keeping this open for now. Don't worry though, it is not forgotten :smile: :rocket: